requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
django-resized==1.0.2
djangorestframework==3.15.2
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
pycodestyle==2.12.0
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
